% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateData.R
\name{generate_data}
\alias{generate_data}
\title{Generate normal distribution data with specified sample number, mean and standard deviation.}
\usage{
generate_data(n, mean, sd)
}
\arguments{
\item{n}{Sample counts.}

\item{mean}{Wanted mean.}

\item{sd}{Wanted standard deviation.}
}
\value{
vector of analog data.
}
\description{
Generate normal distribution data with specified sample number, mean and standard deviation.
}
\examples{
n=10
mean=5.20
sd=0.05
data1 <- rnorm(n, mean, sd)
data1
mean(data1)
sd(data1)
data2 <- generate_data(10, 5.20, 0.05)
data2
mean(data2)
sd(data2)
}
